Disable JFR recording when not supported by the JDK
